fix: medium findings — TS validation, NAPI comment, server log level, docs
- TS: add Number.isInteger guards to maxAssetsPerDeploy and
  maxDeploySizeBytes in validateSiteConfig (rejects 1.5, NaN, Infinity)
- TS: simplify evaluateInvitation — replace manual Promise/string
  branching with await Promise.resolve(raw)
- TS: narrow evaluateInvitation bridge return type (remove bare string)
- NAPI: fix misleading routing ID comment — both paths use SHA-256
- FFI: change tracing::error! to tracing::debug! in server_err/node_err
  across all 3 bridges (PyO3, NAPI, UniFFI) — server operation errors
  are expected at the bridge boundary and should not pollute error logs
- Docs: fix ADR phase-6 attestation timestamp unit (unix_ms → unix_s)
- Swift: change maxDeploySizeBytes from Int to Int64 to avoid 32-bit
  overflow on platforms where Int is 32 bits

+++ bindings/swift/Sources/SCP/Governance.swift
@@ -1585,7 +1585,7 @@ public struct SiteConfig: Sendable, Equatable {
     public let maxAssetsPerDeploy: Int
 
     /// Maximum total deploy size in bytes (default: 536,870,912 = 512 MiB).
-    public let maxDeploySizeBytes: Int
+    public let maxDeploySizeBytes: Int64
 
     /// Number of deploys to retain (default: 2, max 8).
     public let deployRetentionCount: Int
@@ -1608,7 +1608,7 @@ public struct SiteConfig: Sendable, Equatable {
         hostname: String,
         indexPath: String = "/index.html",
         maxAssetsPerDeploy: Int = 10000,
-        maxDeploySizeBytes: Int = 536_870_912,
+        maxDeploySizeBytes: Int64 = 536_870_912,
         deployRetentionCount: Int = 2,
         cspOverride: String? = nil
     ) throws {
